services/semantic/requirements.txt
fastapi
uvicorn
pymorphy2
nltk
sentence-transformers
transliterate
requests
psycopg2-binary
